separate util tests
